fix(gcs-upload): gsutil cp -I cannot keep directory structure
- There are two problems with this approach

1) `gsutil -m cp -I` results in a flat list of files, regardless of directory structure of the input paths.

2) as a consequence of 1), files that have the same name in different directories are overwritten, resulting in a single blob.

To get around 2) I create symlinks for same-named files with an added index-based suffix.

To get around 1) I rename the blobs to get back the input directory structure after everything is uploaded. This is slow since renaming a blob involves a full data copy and delete. This could possibly be done within each dask worker after a chunk is uploaded.